Begin breaking apart default behavior from progressive
