+++ package.json
@@ -41,5 +41,11 @@
     "babel-preset-es2015": "^6.6.0",
     "jest": "^18.0.0",
     "redux-persist": "^4.0.0"
+  },
+  "jest": {
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "<rootDir>/lib/"
+    ]
   }
 }
